requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
Django==3.2.7
django-common-helpers==0.9.2
django-cron==0.5.1
django-redis==5.0.0
djangorestframework==3.12.4
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
pyshorteners==1.0.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
schedule==1.1.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
Werkzeug==2.0.1
